Provisioning: separate contracts and focused regression suites
